+++ src/PoViEmu.Core/Risc/InstTool.cs
@@ -103,7 +103,7 @@ namespace PoViEmu.Core.Risc
         public static BaseOperand N(byte low) => new DestReg(GetReg(low));
         public static BaseOperand M(byte low) => new SourceReg(GetReg(low));
         public static BaseOperand I(byte low) => new ImmedUOperand(low);
-        public static BaseOperand D(ushort low) => new DisplOperand(low);
+        public static BaseOperand D(uint low) => new DisplOperand(low);
         
         public static BaseOperand Nm(byte low) => new MemoryOperand(GetReg(low), IsMinus: true);
         public static BaseOperand M(byte low, ushort dis) => new MemoryOperand(GetReg(low), Dis: dis);
